+++ R/med_mod_code.R
@@ -1,47 +1,52 @@
 #' Conduct (parallel) mediation analysis
 #'
-#' Runs mediation analysis with one or more parallel mediators (using lavaan). The
-#' results can then be plotted with `plot_mediation()`.
+#' Runs mediation analysis with one or more parallel mediators (using the `lavaan`
+#' package). The results can then be plotted with `plot_mediation()`.
 #' 
 #' Note that covariates (if given) are used in predicting each mediator and the outcome. The
 #' coefficients for covariates are returned as an attribute to the main results,
-#' given that they are not always reported. To access them, use 
-#' `attr(res, "CV_coefficients")` if you have saved the return of this function in `res`
+#' given that they are not always reported. To access them, use `attr(res, "CV_coefficients")` 
+#' if you have saved the return of this function in `res`. The lavaan code to estimate the model is
+#' also returned as an attribute, access it with `attr(res, "lavaan_code")`. 
 #'
-#' @param data Dataframe 
-#' @param X Predictor variable (all variables can be passed as character or 'bare')
+#' @param data Data frame 
+#' @param X Predictor variable (all variables should be passed 'bare' in tidyverse style)
 #' @param Y Outcome variable 
 #' @param Ms Mediator variable(s)
-#' @param CVs Covariates (in predicting mediators and outcomes) 
-#' @param standardized_all Should all coefficients (paths, direct and indirect effects) be standardised
-#' @param conf.level The confidence level to be used for confidence intervals. Must be between 0 and 1 
-#' (exclusive), defaults to .95, which corresponds to 95% confidence intervals.
-#' @param seed Random seed, set to get reproducible results. If you do not want to set a fixed seed,
-#' you can use seed = sample(1:1e6, 1)
+#' @param CVs Covariates (in predicting mediators *and* outcomes) 
+#' @param standardized_all Logical. Should all coefficients (paths, direct and indirect effects) be standardized?
+#' @param conf_level The confidence level to be used for confidence intervals. Must be between 0 and 1 
+#' (exclusive), defaults to .95 (i.e. 95% confidence intervals).
+#' @param seed Random seed. You should set this to get reproducible results. 
 #' @param bootstraps Number of bootstraps, defaults to 5000.
 #' @param ... Options passed on to [lavaan::sem()]. 
 #' @return Tibble with direct, total and indirect effects, based on bootstrap resamples. In addition, 'a' coefficients for paths from 
-#' X to mediators and 'b' coefficients for paths from mediators to Y are returned. Coefficients for CVs
-#' are returned as an attribute - see below.
+#' X to mediators and 'b' coefficients for paths from mediators to Y are returned. Coefficients for covariates
+#' are returned as an attribute - see the example.
 #' @export
 #' @examples
-#' # Might link between depression and self-reported health be partly explained
+#' # Might the link between depression and self-reported health be partly explained
 #' # by reductions in physical activity level, when holding age constant?
 #'  
-#'  res <- run_mediation(ess_health, fltdpr, health, dosprt, agea, bootstraps = 100) 
+#'  res <- run_mediation(ess_health, fltdpr, health, dosprt, agea, bootstraps = 50) 
 #' 
 #'  res
 #'  
 #'  attr(res, "CV_coefficients")
 #'  
-#' # NB: bootstraps = 100 only set to reduce running time - should be 1000+
+#' # NB: bootstraps = 50 only set to reduce running time - should be 1000+
 
 run_mediation <- function(data, X, Y, Ms, CVs = NULL, standardized_all = TRUE,
-                          conf.level = .95, seed = 987654321,
+                          conf_level = .95, seed = NULL,
                           bootstraps = 5000, ...) {
 
   .check_req_packages(c("lavaan"))
 
+  args <- as.list(match.call(sys.function(1), sys.call(1), expand.dots = TRUE))[-1]
+  if ("conf.level" %in% names(args)) {
+    stop("The confidence level needs to be specified as conf_level, NOT conf.level")
+  }
+    
   # Convert arguments to strings
   X <- as.character(rlang::ensym(X))
   Y <- as.character(rlang::ensym(Y))
@@ -125,7 +130,8 @@ run_mediation <- function(data, X, Y, Ms, CVs = NULL, standardized_all = TRUE,
   bs <- lavaan::bootstrapLavaan(fit, R = bootstraps, FUN = "coef", parallel = "snow") %>%
     data.frame()
 
-  bs_CVs <- dplyr::select(bs, dplyr::matches(paste0(paste0("\\.", CVs_string), collapse = "|")), -dplyr::matches("\\.\\.")) %>%
+  bs_CVs <- bs %>% 
+    dplyr::select(dplyr::matches(paste0(paste0("\\.", CVs_string), collapse = "|")), -dplyr::matches("\\.\\.")) %>%
     t() %>%
     data.frame()
 
@@ -134,17 +140,19 @@ run_mediation <- function(data, X, Y, Ms, CVs = NULL, standardized_all = TRUE,
     dplyr::rename(M_letter %>% magrittr::set_names(paste0("a__", M_letter))) %>%
     dplyr::rename(paste0(M_letter, M_letter) %>% magrittr::set_names(paste0("b__", M_letter)))
 
+  # Execute dplyr::mutate command based on character string
   char_mutate <- function(data, s) {
     q <- quote(dplyr::mutate(data, z = s))
     eval(parse(text = sub("z = s", s, deparse(q))))
   }
 
-  purrr::map(M_letter, function(x) glue::glue("indirect__{x} = a__{x}*b__{x}")) %>% purrr::walk(function(x) {
+  purrr::map(M_letter, function(x) glue::glue("indirect__{x} = a__{x}*b__{x}")) %>% 
+    purrr::walk(function(x) {
     bs <<- char_mutate(bs, x)
   })
 
-  bs <- char_mutate(bs, paste0("total = cdash + ", paste0("indirect__", M_letter, collapse = " + "))) %>% dplyr::rename(direct = .data$cdash)
-
+  bs <- bs %>% char_mutate(paste0("total = cdash + ", paste0("indirect__", M_letter, collapse = " + "))) %>% 
+    dplyr::rename(direct = .data$cdash)
 
   res <- bs %>%
     t() %>%
@@ -152,7 +160,10 @@ run_mediation <- function(data, X, Y, Ms, CVs = NULL, standardized_all = TRUE,
     tibble::rownames_to_column("parameter") %>%
     tidyr::gather(-.data$parameter, key = "rep", value = "coef") %>%
     dplyr::group_by(.data$parameter) %>%
-    dplyr::summarise(est = mean(.data$coef), se = sd(.data$coef), pvalue = ifelse(.data$est > 0, mean(.data$coef < 0) * 2, mean(.data$coef > 0)) * 2, ci.lower = quantile(.data$coef, (1 - conf.level) / 2), ci.upper = quantile(.data$coef, 1 - (1 - conf.level) / 2)) %>%
+    dplyr::summarise(est = mean(.data$coef), se = sd(.data$coef), 
+                     pvalue = ifelse(.data$est > 0, mean(.data$coef < 0) * 2, mean(.data$coef > 0) * 2), 
+                     ci.lower = quantile(.data$coef, (1 - conf_level) / 2), 
+                     ci.upper = quantile(.data$coef, 1 - (1 - conf_level) / 2)) %>%
     tidyr::separate(.data$parameter, c("type", "mediator"), fill = "right", sep = "__") %>%
     dplyr::mutate(mediator = stringr::str_replace_all(.data$mediator, Ms %>% magrittr::set_names(M_letter)))
 
@@ -160,10 +171,14 @@ run_mediation <- function(data, X, Y, Ms, CVs = NULL, standardized_all = TRUE,
     tibble::rownames_to_column("parameter") %>%
     tidyr::gather(-.data$parameter, key = "rep", value = "coef") %>%
     dplyr::group_by(.data$parameter) %>%
-    dplyr::summarise(est = mean(.data$coef), se = sd(.data$coef), pvalue = ifelse(.data$est > 0, mean(.data$coef < 0) * 2, mean(.data$coef > 0)) * 2, ci.lower = quantile(.data$coef, (1 - conf.level) / 2), ci.upper = quantile(.data$coef, 1 - (1 - conf.level) / 2)) %>%
+    dplyr::summarise(est = mean(.data$coef), se = sd(.data$coef), 
+                     pvalue = ifelse(.data$est > 0, mean(.data$coef < 0) * 2, mean(.data$coef > 0) * 2), 
+                     ci.lower = quantile(.data$coef, (1 - conf_level) / 2), 
+                     ci.upper = quantile(.data$coef, 1 - (1 - conf_level) / 2)) %>%
     tidyr::separate(.data$parameter, c("DV", "CV"), fill = "right", sep = "~")
 
   attr(res, "CV_coefficients") <- CV_res
+  attr(res, "lavaan_code") <- mod
 
   res
 }