+++ R/pcor.R
@@ -1,0 +1,26 @@
+#' Calculates a partial correlation matrix controlling for one variable
+#'
+#' This returns a matrix aligned with the `cor_matrix()` function after parcelling
+#' out the effect of *one* other variable.
+#' 
+#' @inheritParams cor_matrix 
+#' @param given A variable in data. It will be removed from all other variables in data,
+#' before the correlation table is calculated.
+#' @inheritDotParams cor_matrix
+#' @export 
+
+pcor_matrix <- function(data, given, ...) {
+  
+  data %<>% dplyr::select_if(is.numeric)
+  
+  given <- rlang::enexpr(given)
+  
+  data <- data %>% dplyr::mutate(dplyr::across(c(dplyr::everything(), -rlang::as_name(given)),
+                                       ~lm(.x ~ !!given) %>% resid)) %>% 
+    select(-rlang::as_name(given))
+  
+  class(data) <- c("resid_df", class(data))
+  
+  cor_matrix(data, ...)
+  
+}